feat: creating a good project structure
