+++ src/aux/auxiliar_functions.py
