+++ src/main.py
@@ -1,0 +1,1 @@
+from .plot.generate_graphs import *
